Fixed transform errors leading to [bug 3](./bugs/003_cone.png)
Ultimately this came down to a few bugs:
 - Computing of inverses was wrong
 - Our coordinate axes weren't getting flipped correctly
 - Vectors and Points transform differently

The last one is the most critical one:  when multiplying points by a
transform, they are susceptible to *translation*, whereas vectors are
not.  So there were a few places where we were using the wrong types,
and skewing things incorrectly.

Now there's a new bug that prevents me from positioning spheres
correctly.