Implement materials, and substantially refactor Surface_Interaction
Surface_Interaction has always been slightly awkward; it's "construct and fill in" nature in the reference implementation doesn't work so well in Rust.

In order to thread through the emissive properties and material properties correctly, this splits Intersection properties into their own struct (formerly InteractionCommon), and passes that to the places that need intersection data.